refactor(composition): pure move — no shims, module-qualified consumers, snapshot updated
Addresses all four review findings on #6177:

- Remove the three old-path `pub use ironclaw_runner::...` shims from
  composition's root (CodeRabbit/type-placement: relocations update
  consumers, never leave path-preservation re-exports) — EXCEPT one
  documented facade re-export of `reborn_failure_summary_for_category`
  for `ironclaw_reborn_cli`, which consumes composition as its facade and
  must not grow a direct runner edge for one helper (the sanctioned
  composition re-export shape, consumer named in the doc-comment).
- Remove the flattened aliases from ironclaw_runner's root (Gemini/
  IronLoop: the crate documents a directory-of-modules API); all intra-
  crate and external references use module-qualified paths.
- Update docs/plans/composition-pubuse.snapshot for the shrunken surface
  (IronLoop) — verified by composition_public_pub_use_surface_matches_snapshot.
- Rebase onto current main (35 commits): rename detection carried main's
  evolved failure_summary API into the moved file; dropped the PR-era
  auth_prompt import that main has since reverted (#6166).

Verified: architecture suite 62 green (incl. snapshot + re-export-wall
tests); clippy -D warnings clean on runner/composition/cli in both the
all-features and default lanes; root e2e compiles against runner paths.
