+++ crates/ironclaw_reborn_composition/src/projection/turn_events.rs
@@ -29,9 +29,6 @@ use ironclaw_turns::{
 use tokio::sync::{Mutex, OnceCell, Semaphore};
 
 use crate::AuthChallengeProvider;
-use crate::product_auth::api::auth_prompt::{
-    BlockedAuthPromptRequest, auth_prompt_view_for_blocked_auth,
-};
 use ironclaw_runner::failure_summary::{
     pinned_failure_summary_for_category, reborn_failure_summary_for_category_and_detail,
 };
